refactor(service): split aggregator/service.rs into focused submodules
Break the 3924-line src/aggregator/service.rs into a directory module.
AciService stays defined in mod.rs (so child modules keep private-field
access); its methods are split into sibling `impl AciService` blocks:

- mod.rs         — AciService struct, constructors, getters, module wiring
- errors.rs      — ServiceError / UpstreamVerificationError / E2eeError
- config.rs      — AciServiceConfig, ReceiptOwner, provenance validation
- receipt_store.rs — ReceiptStore trait + in-memory store
- clock.rs       — Clock / SystemClock / FixedClock
- wire.rs        — request/response value + middleware/streaming types
- e2ee.rs        — attestation + E2EE request-preparation methods
- e2ee_crypto.rs — E2EE field decrypt/encrypt/AAD helpers
- streaming.rs   — SSE finalizing streams + chat-id parser
- forward.rs     — the buffered/middleware/streaming forward hot paths
- claims.rs      — UpstreamSessionSink + per-provider claim mappers (+ tests)
- receipts.rs    — receipt lookups + legacy signatures
- helpers.rs     — id/body/model small helpers

Pure relocation: no logic changes. Public paths preserved via re-export;
cross-module items marked pub(super). cargo test --all-targets:
220 passed, 0 failed (unchanged from baseline).
