+++ src/aggregator/service/clock.rs
@@ -1,0 +1,25 @@
+/// fixed clock; production uses [`SystemClock`].
+pub trait Clock: Send + Sync {
+    fn now_secs(&self) -> u64;
+}
+
+#[derive(Default)]
+pub struct SystemClock;
+
+impl Clock for SystemClock {
+    fn now_secs(&self) -> u64 {
+        use std::time::{SystemTime, UNIX_EPOCH};
+        SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .map(|d| d.as_secs())
+            .unwrap_or(0)
+    }
+}
+
+pub struct FixedClock(pub u64);
+
+impl Clock for FixedClock {
+    fn now_secs(&self) -> u64 {
+        self.0
+    }
+}
